Add support for Symfony 4/5 and remove Symfony 2 support
This adds support for both Symfony 4 & 5, while removing support for
Symfony 2.

The command provided by this package needed to be explicitly exposed as
a public service for Symfony 4.

`ContainerAwareCommand` is removed in Symfony 5. Instead the container
is injected into the command so it is available for use by the Boris
repl.